Don't pass the hostgroup to the run_puppet command
It does not make sense to pass both zone and hostgroup to cicd.run_puppet even
more so when no node is specified. Indeed it would set the '--environment' in a
unexpected way.

Setting the '--environment' is only useful for the first run puppet and it
shouldn't be done across nodes anyway.

We still need to support a first run_puppet but setting '--environment' should
be explicit and not part of the targeting.